+++ src/core/backup/index.ts
@@ -1,3 +1,5 @@
 export * from "./BackupService";
 export * from "./BackupStorage";
+export * from "./backupMigration";
+export * from "./backupValidation";
 export * from "./types";